refactor(docs): rename boot void to particle orbit
