+++ apps/docs/examples/particle-orbit/blur.wgsl
