+++ apps/docs/examples/particle-orbit/bright.wgsl
