+++ apps/docs/examples/particle-orbit/jfa-init.wgsl
