+++ apps/docs/examples/particle-orbit/rc-emitter.wgsl
