+++ apps/docs/examples/particle-orbit/rc-resolve.wgsl
