+++ apps/docs/examples/particle-orbit/render-thumbnail.ts
