+++ apps/docs/examples/particle-orbit/stars.wgsl
