+++ apps/docs/lib/example-components.ts
@@ -31,7 +31,7 @@ export const exampleComponentLoaders = {
   'mnist-classifier': () => import('../examples/mnist-classifier/index'),
   'air-painting': () => import('../examples/air-painting/index'),
   'three-tsl': () => import('../examples/three-tsl/index'),
-  'boot-void': () => import('../examples/boot-void/index'),
+  'particle-orbit': () => import('../examples/particle-orbit/index'),
 } satisfies Record<ExampleSlug, ExampleComponentLoader>;
 
 export function getExampleComponentLoader(slug: ExampleSlug): ExampleComponentLoader {
