+++ apps/docs/lib/example-slugs.ts
@@ -22,7 +22,7 @@ export const exampleSlugs = [
   'mnist-classifier',
   'air-painting',
   'three-tsl',
-  'boot-void',
+  'particle-orbit',
 ] as const;
 
 export type ExampleSlug = (typeof exampleSlugs)[number];
