+++ apps/docs/lib/examples-source.generated.ts
@@ -1714,10 +1714,10 @@ export const exampleSources = {
       }
     ]
   },
-  "boot-void": {
-    "slug": "boot-void",
-    "title": "Boot Void",
-    "description": "A retro console boot homage co-driven by vgpu and TypeGPU on one shared device: TypeGPU simulates light-seeking dust with typed schemas, fed by the irradiance vgpu renders through radiance cascades, and vgpu draws the particle buffer back zero-copy. Plus a render bundle, bloom, and a CRT finish.",
+  "particle-orbit": {
+    "slug": "particle-orbit",
+    "title": "Particle Orbit",
+    "description": "A shared-device vgpu + TypeGPU experiment: TypeGPU advances light-seeking dust through a radiance field rendered by vgpu, then vgpu draws the same particle buffer zero-copy alongside orbiting lights, HDR bloom, and a CRT finish.",
     "tags": [
       "instancing",
       "particles",
@@ -1760,7 +1760,7 @@ export const exampleSources = {
       {
         "path": "pipeline.ts",
         "language": "typescript",
-        "content": "import {\n  bundle,\n  draw,\n  effect,\n  sampler,\n  target,\n  type Bundle,\n  type Frame,\n  type Gpu,\n  type Surface,\n  type Target,\n} from 'vgpu';\n\nimport atmosphereWgsl from './atmosphere.wgsl';\nimport {\n  encodeRadiance,\n  setRadianceScene,\n  setRadianceTime,\n  type Radiance,\n} from './radiance';\nimport blurWgsl from './blur.wgsl';\nimport brightWgsl from './bright.wgsl';\nimport nebulaWgsl from './nebula.wgsl';\nimport postWgsl from './post.wgsl';\nimport starsWgsl from './stars.wgsl';\nimport trailsWgsl from './trails.wgsl';\n\ntype Output = Surface | Target;\n\nconst CLEAR = [0, 0, 0, 1] as const;\nconst SCENE_FORMAT = 'rgba16float' as const;\n// 4 trails × (191 history segments + 1 head halo).\nexport const TRAIL_INSTANCES = 768;\n// Dense enough for each orbiting light to tint several motes as it passes.\nexport const DUST_COUNT = 1024;\n\nconst BLURS = [\n  { direction: [1, 0], radius: 1 },\n  { direction: [0, 1], radius: 1 },\n  { direction: [1, 0], radius: 2.4 },\n  { direction: [0, 1], radius: 2.4 },\n] as const;\n\nexport function createEffects(gpu: Gpu) {\n  const samp = sampler(gpu, { minFilter: 'linear', magFilter: 'linear' });\n  const additive = {\n    color: { src: 'src-alpha', dst: 'one' },\n    alpha: { src: 'one', dst: 'one' },\n  } as const;\n  return {\n    nebula: effect(gpu, nebulaWgsl),\n    stars: draw(gpu, {\n      shader: starsWgsl,\n      vertices: 6,\n      instances: DUST_COUNT,\n      blend: additive,\n      label: 'boot-void-stars',\n    }),\n    atmosphere: effect(gpu, atmosphereWgsl, { set: { samp } }),\n    trails: draw(gpu, {\n      shader: trailsWgsl,\n      vertices: 6,\n      instances: TRAIL_INSTANCES,\n      blend: additive,\n      label: 'boot-void-trails',\n    }),\n    bright: effect(gpu, brightWgsl, { set: { samp } }),\n    blur: BLURS.map((options) => effect(gpu, blurWgsl, { set: { samp, blur: options } })),\n    post: effect(gpu, postWgsl, { set: { samp, params: { pointer: [0, 0] } } }),\n  };\n}\n\nexport type Effects = ReturnType<typeof createEffects>;\n\nexport function createTargets(gpu: Gpu, size: readonly [number, number]) {\n  const owned: Target[] = [];\n  const own = (created: Target) => {\n    owned.push(created);\n    return created;\n  };\n  try {\n    const bloomHeight = Math.min(360, size[1]);\n    const bloomSize: [number, number] = [\n      Math.max(1, Math.round((bloomHeight * size[0]) / size[1])),\n      bloomHeight,\n    ];\n    return {\n      // 4× MSAA: passes resolve into a sampleable single-sample color.\n      scene: own(target(gpu, { size, format: SCENE_FORMAT, msaa: true })),\n      composite: own(target(gpu, { size, format: SCENE_FORMAT })),\n      bloom: [\n        own(target(gpu, { size: bloomSize, format: SCENE_FORMAT })),\n        own(target(gpu, { size: bloomSize, format: SCENE_FORMAT })),\n      ] as const,\n    };\n  } catch (error) {\n    for (const created of owned.reverse()) destroy(created);\n    throw error;\n  }\n}\n\nexport type Targets = ReturnType<typeof createTargets>;\n\nexport function destroyTargets(targets: Targets): void {\n  destroy(targets.bloom[1]);\n  destroy(targets.bloom[0]);\n  destroy(targets.composite);\n  destroy(targets.scene);\n}\n\nfunction destroy(color: Target | undefined): void {\n  (color as { destroy?: () => void } | undefined)?.destroy?.();\n}\n\nexport function setBindings(effects: Effects, targets: Targets, radiance: Radiance): void {\n  const aspect = targets.scene.size[0] / targets.scene.size[1];\n  effects.nebula.set({ params: { aspect } });\n  effects.stars.set({ params: { aspect } });\n  effects.atmosphere.set({ src: targets.scene, params: { aspect } });\n  effects.trails.set({ params: { aspect } });\n  effects.bright.set({ src: targets.composite });\n  effects.blur.forEach((pass, i) =>\n    pass.set({\n      src: targets.bloom[i % 2],\n      blur: { texelSize: targets.bloom[i % 2].texelSize },\n    }),\n  );\n  effects.post.set({\n    src: targets.composite,\n    bloom: targets.bloom[0],\n    radiance: radiance.irradiance,\n    params: { aspect },\n  });\n  setRadianceScene(radiance, targets.scene);\n}\n\nexport function setTime(effects: Effects, radiance: Radiance, time: number): void {\n  const params = { params: { time } };\n  effects.nebula.set(params);\n  effects.stars.set(params);\n  effects.atmosphere.set(params);\n  effects.trails.set(params);\n  effects.post.set(params);\n  setRadianceTime(radiance, time);\n}\n\nexport function setPointer(effects: Effects, pointer: readonly [number, number]): void {\n  effects.post.set({ params: { pointer } });\n}\n\nexport async function prewarm(effects: Effects, targets: Targets, output: Output): Promise<void> {\n  await Promise.all([\n    effects.nebula.compile(targets.scene),\n    effects.stars.compile(targets.scene),\n    effects.atmosphere.compile(targets.composite),\n    effects.trails.compile(targets.composite),\n    effects.bright.compile(targets.bloom[0]),\n    ...effects.blur.map((pass, i) => pass.compile(targets.bloom[(i + 1) % 2])),\n    effects.post.compile({ colors: [output.format] }),\n  ]);\n}\n\n// Bundles freeze the render signature, not the target, so they survive resizes.\nexport function recordScene(gpu: Gpu, effects: Effects): Bundle {\n  return bundle(\n    gpu,\n    { target: { colors: [SCENE_FORMAT], sampleCount: 4 }, label: 'boot-void-scene' },\n    (pass) => {\n      pass.draw(effects.nebula);\n      pass.draw(effects.stars);\n    },\n  );\n}\n\nexport function renderChain(\n  frame: Frame,\n  effects: Effects,\n  targets: Targets,\n  output: Output,\n  scene: Bundle,\n  radiance: Radiance,\n): void {\n  frame.pass({ target: targets.scene, clear: CLEAR }, (pass) => pass.bundles(scene));\n  encodeRadiance(frame, radiance);\n  frame.pass({ target: targets.composite, clear: CLEAR }, (pass) => {\n    pass.draw(effects.atmosphere);\n    pass.draw(effects.trails);\n  });\n  frame.pass({ target: targets.bloom[0], clear: CLEAR }, (pass) => pass.draw(effects.bright));\n  effects.blur.forEach((pass, i) => {\n    frame.pass({ target: targets.bloom[(i + 1) % 2], clear: CLEAR }, (framePass) =>\n      framePass.draw(pass),\n    );\n  });\n  frame.pass({ target: output, clear: CLEAR }, (pass) => pass.draw(effects.post));\n}\n"
+        "content": "import {\n  bundle,\n  draw,\n  effect,\n  sampler,\n  target,\n  type Bundle,\n  type Frame,\n  type Gpu,\n  type Surface,\n  type Target,\n} from 'vgpu';\n\nimport atmosphereWgsl from './atmosphere.wgsl';\nimport {\n  encodeRadiance,\n  setRadianceScene,\n  setRadianceTime,\n  type Radiance,\n} from './radiance';\nimport blurWgsl from './blur.wgsl';\nimport brightWgsl from './bright.wgsl';\nimport nebulaWgsl from './nebula.wgsl';\nimport postWgsl from './post.wgsl';\nimport starsWgsl from './stars.wgsl';\nimport trailsWgsl from './trails.wgsl';\n\ntype Output = Surface | Target;\n\nconst CLEAR = [0, 0, 0, 1] as const;\nconst SCENE_FORMAT = 'rgba16float' as const;\n// 4 trails × (191 history segments + 1 head halo).\nexport const TRAIL_INSTANCES = 768;\n// Dense enough for each orbiting light to tint several motes as it passes.\nexport const DUST_COUNT = 1024;\n\nconst BLURS = [\n  { direction: [1, 0], radius: 1 },\n  { direction: [0, 1], radius: 1 },\n  { direction: [1, 0], radius: 2.4 },\n  { direction: [0, 1], radius: 2.4 },\n] as const;\n\nexport function createEffects(gpu: Gpu) {\n  const samp = sampler(gpu, { minFilter: 'linear', magFilter: 'linear' });\n  const additive = {\n    color: { src: 'src-alpha', dst: 'one' },\n    alpha: { src: 'one', dst: 'one' },\n  } as const;\n  return {\n    nebula: effect(gpu, nebulaWgsl),\n    stars: draw(gpu, {\n      shader: starsWgsl,\n      vertices: 6,\n      instances: DUST_COUNT,\n      blend: additive,\n      label: 'particle-orbit-stars',\n    }),\n    atmosphere: effect(gpu, atmosphereWgsl, { set: { samp } }),\n    trails: draw(gpu, {\n      shader: trailsWgsl,\n      vertices: 6,\n      instances: TRAIL_INSTANCES,\n      blend: additive,\n      label: 'particle-orbit-trails',\n    }),\n    bright: effect(gpu, brightWgsl, { set: { samp } }),\n    blur: BLURS.map((options) => effect(gpu, blurWgsl, { set: { samp, blur: options } })),\n    post: effect(gpu, postWgsl, { set: { samp, params: { pointer: [0, 0] } } }),\n  };\n}\n\nexport type Effects = ReturnType<typeof createEffects>;\n\nexport function createTargets(gpu: Gpu, size: readonly [number, number]) {\n  const owned: Target[] = [];\n  const own = (created: Target) => {\n    owned.push(created);\n    return created;\n  };\n  try {\n    const bloomHeight = Math.min(360, size[1]);\n    const bloomSize: [number, number] = [\n      Math.max(1, Math.round((bloomHeight * size[0]) / size[1])),\n      bloomHeight,\n    ];\n    return {\n      // 4× MSAA: passes resolve into a sampleable single-sample color.\n      scene: own(target(gpu, { size, format: SCENE_FORMAT, msaa: true })),\n      composite: own(target(gpu, { size, format: SCENE_FORMAT })),\n      bloom: [\n        own(target(gpu, { size: bloomSize, format: SCENE_FORMAT })),\n        own(target(gpu, { size: bloomSize, format: SCENE_FORMAT })),\n      ] as const,\n    };\n  } catch (error) {\n    for (const created of owned.reverse()) destroy(created);\n    throw error;\n  }\n}\n\nexport type Targets = ReturnType<typeof createTargets>;\n\nexport function destroyTargets(targets: Targets): void {\n  destroy(targets.bloom[1]);\n  destroy(targets.bloom[0]);\n  destroy(targets.composite);\n  destroy(targets.scene);\n}\n\nfunction destroy(color: Target | undefined): void {\n  (color as { destroy?: () => void } | undefined)?.destroy?.();\n}\n\nexport function setBindings(effects: Effects, targets: Targets, radiance: Radiance): void {\n  const aspect = targets.scene.size[0] / targets.scene.size[1];\n  effects.nebula.set({ params: { aspect } });\n  effects.stars.set({ params: { aspect } });\n  effects.atmosphere.set({ src: targets.scene, params: { aspect } });\n  effects.trails.set({ params: { aspect } });\n  effects.bright.set({ src: targets.composite });\n  effects.blur.forEach((pass, i) =>\n    pass.set({\n      src: targets.bloom[i % 2],\n      blur: { texelSize: targets.bloom[i % 2].texelSize },\n    }),\n  );\n  effects.post.set({\n    src: targets.composite,\n    bloom: targets.bloom[0],\n    radiance: radiance.irradiance,\n    params: { aspect },\n  });\n  setRadianceScene(radiance, targets.scene);\n}\n\nexport function setTime(effects: Effects, radiance: Radiance, time: number): void {\n  const params = { params: { time } };\n  effects.nebula.set(params);\n  effects.stars.set(params);\n  effects.atmosphere.set(params);\n  effects.trails.set(params);\n  effects.post.set(params);\n  setRadianceTime(radiance, time);\n}\n\nexport function setPointer(effects: Effects, pointer: readonly [number, number]): void {\n  effects.post.set({ params: { pointer } });\n}\n\nexport async function prewarm(effects: Effects, targets: Targets, output: Output): Promise<void> {\n  await Promise.all([\n    effects.nebula.compile(targets.scene),\n    effects.stars.compile(targets.scene),\n    effects.atmosphere.compile(targets.composite),\n    effects.trails.compile(targets.composite),\n    effects.bright.compile(targets.bloom[0]),\n    ...effects.blur.map((pass, i) => pass.compile(targets.bloom[(i + 1) % 2])),\n    effects.post.compile({ colors: [output.format] }),\n  ]);\n}\n\n// Bundles freeze the render signature, not the target, so they survive resizes.\nexport function recordScene(gpu: Gpu, effects: Effects): Bundle {\n  return bundle(\n    gpu,\n    { target: { colors: [SCENE_FORMAT], sampleCount: 4 }, label: 'particle-orbit-scene' },\n    (pass) => {\n      pass.draw(effects.nebula);\n      pass.draw(effects.stars);\n    },\n  );\n}\n\nexport function renderChain(\n  frame: Frame,\n  effects: Effects,\n  targets: Targets,\n  output: Output,\n  scene: Bundle,\n  radiance: Radiance,\n): void {\n  frame.pass({ target: targets.scene, clear: CLEAR }, (pass) => pass.bundles(scene));\n  encodeRadiance(frame, radiance);\n  frame.pass({ target: targets.composite, clear: CLEAR }, (pass) => {\n    pass.draw(effects.atmosphere);\n    pass.draw(effects.trails);\n  });\n  frame.pass({ target: targets.bloom[0], clear: CLEAR }, (pass) => pass.draw(effects.bright));\n  effects.blur.forEach((pass, i) => {\n    frame.pass({ target: targets.bloom[(i + 1) % 2], clear: CLEAR }, (framePass) =>\n      framePass.draw(pass),\n    );\n  });\n  frame.pass({ target: output, clear: CLEAR }, (pass) => pass.draw(effects.post));\n}\n"
       },
       {
         "path": "dust.ts",
