+++ apps/docs/scripts/example-chunk-budgets.json
@@ -30,7 +30,7 @@
     "mnist-classifier": 63571,
     "air-painting": 71822,
     "three-tsl": 247268,
-    "boot-void": 160967
+    "particle-orbit": 160967
   },
   "exampleGrowth": {
     "percent": 2,
